enh: united all changes and ordered them
