+++ src/utils/httprequest.js
@@ -1,14 +1,12 @@
 const https = require("https");
 const fs = require("fs");
-const ora = require("ora");
-const chalk = require("chalk");
-const { movie_render } = require("./movie_render");
-const spinner = require("./spinner_persons");
-const render = require("./render_persons");
+const { movie_render } = require("./render_movie");
+const spin = require("./spinner");
+const render = require("./render_globals");
 require("dotenv/config");
 
 exports.httpRequest = function (endPoint, option1 = "", option2 = "") {
-  const spinner = ora("Loading Film").start();
+  const spinner = spin.start(endPoint);
   https
     .get(
       `https://api.themoviedb.org/3/${endPoint}?api_key=${process.env.API_KEY}&${option1}&${option2}`,
@@ -20,27 +18,26 @@ exports.httpRequest = function (endPoint, option1 = "", option2 = "") {
 
         response.on("end", () => {
           if (result === "") {
-            spin.fail("Error: can't find your request");
+            spinner.fail("Error: can't find your request");
             return;
           }
 
-          switch (endPoint) {
-            case "person/popular":
-              render.persons(JSON.parse(result));
-              spin.succeed("Popular Persons data loaded");
-              break;
-            case "movie/popular":
-              render.movies(JSON.parse(result));
-              spin.succeed("Popular movies data loaded");
-              break;
-            case "movie/now_playing":
-              render.movies(JSON.parse(result));
-              spin.succeed("Movies playing now data loaded");
-              break;
-            case "movie/70":
-              movie_render(JSON.parse(result));
-              spinner.succeed("Film loaded successfully");
-              break;
+          if (endPoint === "person/popular") {
+            render.persons(JSON.parse(result));
+            spinner.succeed("Popular Persons data loaded");
+            return;
+          } else if (endPoint === "movie/popular") {
+            render.movies(JSON.parse(result));
+            spinner.succeed("Popular movies data loaded");
+            return;
+          } else if (endPoint === "movie/now_playing") {
+            render.movies(JSON.parse(result));
+            spinner.succeed("Movies playing now data loaded");
+            return;
+          } else if (endPoint.includes("movie/")) {
+            movie_render(JSON.parse(result));
+            spinner.succeed("Film loaded successfully");
+            return;
           }
         });
       }
