+++ src/utils/render_globals.js
